feat(swagger): add BkStandardResponseSwaggerAutoSchema
