+++ src/dashboard/apigateway/apigateway/apis/web/monitor/views.py
@@ -26,7 +26,6 @@ from apigateway.apps.monitor.models import AlarmRecord, AlarmStrategy
 from apigateway.biz.monitor import ResourceMonitorHandler
 from apigateway.common.factories import SchemaFactory
 from apigateway.utils.responses import OKJsonResponse
-from apigateway.utils.swagger import PaginatedResponseSwaggerAutoSchema
 from apigateway.utils.time import now_datetime
 
 from . import filters
@@ -45,7 +44,6 @@ from .serializers import (
 @method_decorator(
     name="get",
     decorator=swagger_auto_schema(
-        auto_schema=PaginatedResponseSwaggerAutoSchema,
         query_serializer=AlarmStrategyQueryInputSLZ,
         responses={status.HTTP_200_OK: AlarmStrategyListOutputSLZ(many=True)},
         tags=["WebAPI.Monitor"],
@@ -199,7 +197,6 @@ class AlarmStrategyUpdateStatusApi(generics.UpdateAPIView):
 @method_decorator(
     name="get",
     decorator=swagger_auto_schema(
-        auto_schema=PaginatedResponseSwaggerAutoSchema,
         query_serializer=AlarmRecordQueryInputSLZ,
         responses={status.HTTP_200_OK: AlarmRecordQueryOutputSLZ(many=True)},
         tags=["WebAPI.Monitor"],
