+++ src/dashboard/apigateway/apigateway/apis/web/permission/views.py
@@ -40,7 +40,6 @@ from apigateway.apps.permission.tasks import send_mail_for_perm_handle
 from apigateway.biz.permission import PermissionDimensionManager
 from apigateway.core.models import Resource
 from apigateway.utils.responses import DownloadableResponse, OKJsonResponse
-from apigateway.utils.swagger import PaginatedResponseSwaggerAutoSchema
 
 from .filters import (
     AppGatewayPermissionFilter,
@@ -74,7 +73,6 @@ class AppResourcePermissionQuerySetMixin:
 @method_decorator(
     name="get",
     decorator=swagger_auto_schema(
-        auto_schema=PaginatedResponseSwaggerAutoSchema,
         responses={status.HTTP_200_OK: AppResourcePermissionOutputSLZ(many=True)},
         tags=["WebAPI.Permission"],
     ),
@@ -258,7 +256,6 @@ class AppGatewayPermissionQuerySetMixin:
 @method_decorator(
     name="get",
     decorator=swagger_auto_schema(
-        auto_schema=PaginatedResponseSwaggerAutoSchema,
         responses={status.HTTP_200_OK: AppGatewayPermissionOutputSLZ(many=True)},
         tags=["WebAPI.Permission"],
     ),
@@ -439,7 +436,6 @@ class AppPermissionApplyQuerySetMixin:
 @method_decorator(
     name="get",
     decorator=swagger_auto_schema(
-        auto_schema=PaginatedResponseSwaggerAutoSchema,
         responses={status.HTTP_200_OK: AppPermissionApplyOutputSLZ(many=True)},
         tags=["WebAPI.Permission"],
     ),
@@ -479,7 +475,6 @@ class AppPermissionApplyRetrieveApi(AppPermissionApplyQuerySetMixin, generics.Re
 @method_decorator(
     name="get",
     decorator=swagger_auto_schema(
-        auto_schema=PaginatedResponseSwaggerAutoSchema,
         responses={status.HTTP_200_OK: AppPermissionRecordOutputSLZ(many=True)},
         tags=["WebAPI.Permission"],
     ),
