+++ src/dashboard/apigateway/apigateway/apis/web/resource/views.py
@@ -42,7 +42,6 @@ from apigateway.common.contexts import ResourceAuthContext
 from apigateway.core.constants import STAGE_VAR_PATTERN
 from apigateway.core.models import Backend, BackendConfig, Proxy, Resource, Stage
 from apigateway.utils.responses import DownloadableResponse, OKJsonResponse
-from apigateway.utils.swagger import PaginatedResponseSwaggerAutoSchema
 
 from .serializers import (
     BackendPathCheckInputSLZ,
@@ -71,7 +70,6 @@ class ResourceQuerySetMixin:
 @method_decorator(
     name="get",
     decorator=swagger_auto_schema(
-        auto_schema=PaginatedResponseSwaggerAutoSchema,
         query_serializer=ResourceQueryInputSLZ,
         responses={status.HTTP_200_OK: ResourceListOutputSLZ(many=True)},
         tags=["WebAPI.Resource"],
